+++ src/Products/PloneMeeting/profiles/__init__.py
@@ -487,8 +487,7 @@ class MeetingConfigDescriptor(Descriptor):
         self.annexToPrintMode = 'enabled_for_info'
         self.keepOriginalToPrintOfClonedItems = True
         self.removeAnnexesPreviewsOnMeetingClosure = False
-        self.cssClassesToHide = 'highlight\nhighlight-red'
-        self.hideCssClassesTo = ()
+        self.cssTransforms = []
         self.enabledItemActions = ('duplication', 'export_pdf')
         self.enabledAnnexesBatchActions = ['download-annexes']